Move make_bitfield_serde to struct_accessors.
Fixes <https://github.com/oxidecomputer/amd-apcb/issues/159>.
